# api-server/src/api/swagger/swaggerJson.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is the Swagger Documentation implementation of API's on NodeJS. You will find detailed documentation and model data about the REST API's. To check if an API is working , execute the API after entering the valid parameters",
    "version": "1.0.0",
    "title": "Parcel App",
    "termsOfService": "",
    "contact": {
      "email": "api@boiler.com"
    }
  },
  "host": "127.0.0.1:4001",
  "basePath": "/api/v1",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/auth/forgot-password": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "forgot password",
        "description": "forgot password",
        "operationId": "forgotPassword",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "forgotPassword",
            "required": true,
            "schema": {
              "$ref": "#/definitions/forgotPassword"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully sent mail"
          },
          "401": {
            "description": "Invalid username"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "forgotPassword": {
      "type": "object",
      "required": [
        "email"
      ],
      "properties": {
        "email": {
          "type": "string",
          "example": "keyurkd07@gmail.com"
        }
      },
      "xml": {
        "name": "forgotPassword"
      }
    },
...
}
}